move Context trait implementation in its own file
